requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
pyyaml==6.0.2
joblib==1.4.2
pytest==8.3.2
ruff==0.6.9
black==24.8.0
isort==5.13.2
mypy==1.11.2
